requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pg8000==1.30.3
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
httpx==0.25.2
pdfplumber==0.10.0
openpyxl==3.1.2
